Changed emm to make it admissible and add some extra heuristics

+++ algorithms/heuristics/combination.py
@@ -2,6 +2,5 @@ from utils.state import SokobanState
 from algorithms.heuristics.deadlock import deadlock_heuristic 
 from algorithms.heuristics.emm import emm_heuristic
 
-
-def combination_heuristic(state: SokobanState) -> int:
+def combination_heuristic(state: SokobanState) -> float:
     return max(deadlock_heuristic(state), emm_heuristic(state))
